Rename mock module
